+++ app/common/src/utilities/data/object.ts
@@ -134,17 +134,31 @@ export function singletonObjectOrNull(value: unknown): [] | [object] {
 // ============
 
 /** UNSAFE when `Ks` contains strings that are not in the runtime array. */
-export function omit<T, Ks extends readonly (string & keyof T)[] | []>(
+export function omit<T, Ks extends readonly [string & keyof T, ...(string & keyof T)[]]>(
   object: T,
   ...keys: Ks
 ): Omit<T, Ks[number]> {
   const keysSet = new Set<string>(keys)
   return Object.fromEntries(
-    // This is SAFE, as it is a reaonly upcast.
-    Object.entries(object as Readonly<Record<string, unknown>>).flatMap(kv =>
-      !keysSet.has(kv[0]) ? [kv] : [],
-    ),
+    // This is SAFE, as it is a readonly upcast.
+    Object.entries(object as Readonly<Record<string, unknown>>).filter(([k]) => !keysSet.has(k)),
   ) as Omit<T, Ks[number]>
+}
+
+// ============
+// === pick ===
+// ============
+
+/** UNSAFE when `Ks` contains strings that are not in the runtime array. */
+export function pick<T, Ks extends readonly [string & keyof T, ...(string & keyof T)[]]>(
+  object: T,
+  ...keys: Ks
+): Pick<T, Ks[number]> {
+  const keysSet = new Set<string>(keys)
+  return Object.fromEntries(
+    // This is SAFE, as it is a readonly upcast.
+    Object.entries(object as Readonly<Record<string, unknown>>).filter(([k]) => keysSet.has(k)),
+  ) as Pick<T, Ks[number]>
 }
 
 // ===================
@@ -162,3 +176,24 @@ export type ExtractKeys<T, U> = {
 
 /** An instance method of the given type. */
 export type MethodOf<T> = (this: T, ...args: never) => unknown
+
+// ===================
+// === useObjectId ===
+// ===================
+
+/** Composable providing support for managing object identities. */
+export function useObjectId() {
+  let lastId = 0
+  const idNumbers = new WeakMap<object, number>()
+  /** @returns A value that can be used to compare object identity. */
+  function objectId(o: object): number {
+    const id = idNumbers.get(o)
+    if (id == null) {
+      lastId += 1
+      idNumbers.set(o, lastId)
+      return lastId
+    }
+    return id
+  }
+  return { objectId }
+}