+++ app/gui/src/project-view/components/GraphEditor/clipboard.ts
@@ -2,9 +2,10 @@ import type { NodeCreationOptions } from '@/composables/nodeCreation'
 import type { GraphStore, Node, NodeId } from '@/stores/graph'
 import { Ast } from '@/util/ast'
 import { Pattern } from '@/util/ast/match'
-import { filterDefined } from '@/util/data/iterable'
+import { nodeDocumentationText } from '@/util/ast/node'
 import { Vec2 } from '@/util/data/vec2'
 import type { ToValue } from '@/util/reactivity'
+import * as iter from 'enso-common/src/utilities/data/iter'
 import { computed, toValue } from 'vue'
 import type { NodeMetadataFields } from 'ydoc-shared/ast'
 
@@ -131,7 +132,7 @@ async function decodeClipboard<T>(
       }
     }
   }
-  return filterDefined(await Promise.all(clipboardItems.map(decodeItem)))
+  return iter.filterDefined(await Promise.all(clipboardItems.map(decodeItem)))
 }
 
 // === Spreadsheet clipboard decoder ===
@@ -186,10 +187,9 @@ export function writeClipboard(data: MimeData) {
 // === Serializing nodes ===
 
 function nodeStructuredData(node: Node): CopiedNode {
-  const documentation = node.outerAst.isStatement() ? node.outerAst.documentationText() : undefined
   return {
     expression: node.innerExpr.code(),
-    documentation,
+    documentation: nodeDocumentationText(node) || undefined,
     metadata: node.rootExpr.serializeMetadata(),
     ...(node.pattern ? { binding: node.pattern.code() } : {}),
   }