+++ app/gui/src/project-view/components/MarkdownEditor/imageUrlTransformer.ts
@@ -1,17 +1,30 @@
 import { createContextStore } from '@/providers'
 import type { ToValue } from '@/util/reactivity'
-import { toValue } from 'vue'
+import { Ref, toValue } from 'vue'
 import { mapOk, Ok, type Result } from 'ydoc-shared/util/data/result'
 
-export type TransformUrlResult = Result<{ url: string; dispose?: () => void }>
+/**
+ * A transformed URL.
+ *
+ * Once the returned URL is not used anymore, `dispose` callback is called allowing release
+ * any resource under that URL.
+ *
+ * `uploading` is set to true while the image is being uploaded to its target destination
+ * (as part of pasting image, for example).
+ */
+export type TransformUrlResult = Result<{
+  url: string
+  dispose?: () => void
+  uploading?: Ref<boolean>
+}>
 export type UrlTransformer = (url: string) => Promise<TransformUrlResult>
 
 export {
-  injectFn as injectLexicalImageUrlTransformer,
-  provideFn as provideLexicalImageUrlTransformer,
+  injectFn as injectDocumentationImageUrlTransformer,
+  provideFn as provideDocumentationImageUrlTransformer,
 }
 const { provideFn, injectFn } = createContextStore(
-  'Lexical image URL transformer',
+  'Documentation image URL transformer',
   (transformUrl: ToValue<UrlTransformer | undefined>) => ({
     transformUrl: (url: string) => toValue(transformUrl)?.(url),
   }),
@@ -22,11 +35,17 @@ type Url = string
 export interface ResourceInfo<T> {
   location: T
   uniqueId: ResourceId
+  uploading?: Ref<boolean>
 }
 export type ResourceLocator<T> = (url: Url) => Promise<Result<ResourceInfo<T>> | undefined>
 export type ResourceFetcher<T> = (locator: T) => Promise<Result<Blob>>
 
-/** TODO: Add docs */
+/**
+ * Create {@link UrlTransformer} which fetches and caches the image. Returns a URL created
+ * with `URL.createObjectURL`.
+ *
+ * May be used in cases, when the image is not available for browser through HTTP protocol.
+ */
 export function fetcherUrlTransformer<ResourceLocation>(
   locateResource: ResourceLocator<ResourceLocation>,
   fetchResource: ResourceFetcher<ResourceLocation>,
@@ -48,7 +67,7 @@ export function fetcherUrlTransformer<ResourceLocation>(
     } else if (!resource.ok) {
       return resource
     } else {
-      const { uniqueId, location } = resource.value
+      const { uniqueId, location, uploading } = resource.value
       const result = await (allocatedUrls.get(uniqueId) ?? startFetch(uniqueId, location))
       if (!result.ok) {
         // Changes to external state may allow a future attempt to succeed.
@@ -59,8 +78,12 @@ export function fetcherUrlTransformer<ResourceLocation>(
       return Ok({
         url: result.value.value,
         dispose: () => {
-          if (!(result.value.refs -= 1)) URL.revokeObjectURL(result.value.value)
+          if (!(result.value.refs -= 1)) {
+            URL.revokeObjectURL(result.value.value)
+            allocatedUrls.delete(uniqueId)
+          }
         },
+        uploading,
       })
     }
   }