+++ app/gui/src/project-view/components/visualizations/tableVizToolbar.ts
@@ -1,10 +1,11 @@
-import { NodeCreationOptions } from '@/components/GraphEditor/nodeCreation'
-import { TextFormatOptions } from '@/components/visualizations/TableVisualization.vue'
-import { ToolbarItem } from '@/components/visualizations/toolbar'
+import type { NodeCreationOptions } from '@/components/GraphEditor/nodeCreation'
+import type { TextFormatOptions } from '@/components/visualizations/TableVisualization.vue'
+import type { ToolbarItem } from '@/components/visualizations/toolbar'
 import { Ast } from '@/util/ast'
 import { Pattern } from '@/util/ast/match'
-import { ToValue } from '@/util/reactivity'
-import { computed, ComputedRef, Ref, toValue } from 'vue'
+import type { ToValue } from '@/util/reactivity'
+import { computed, type ComputedRef, type Ref, toValue } from 'vue'
+import { Expression, MutableExpression } from 'ydoc-shared/ast'
 
 type SortDirection = 'asc' | 'desc'
 export type SortModel = {
@@ -24,6 +25,9 @@ export interface SortFilterNodesButtonOptions {
   isDisabled: ToValue<boolean>
   isFilterSortNodeEnabled: ToValue<boolean>
   createNodes: (...options: NodeCreationOptions[]) => void
+  getColumnValueToEnso: (
+    columnName: string,
+  ) => (columnValue: string, module: Ast.MutableModule) => Ast.Owned<Ast.MutableAst>
 }
 
 export interface FormatMenuOptions {
@@ -38,6 +42,7 @@ function useSortFilterNodesButton({
   isDisabled,
   isFilterSortNodeEnabled,
   createNodes,
+  getColumnValueToEnso,
 }: SortFilterNodesButtonOptions): ComputedRef<ToolbarItem | undefined> {
   const sortPatternPattern = computed(() => Pattern.parseExpression('(..Name __ __ )')!)
 
@@ -73,7 +78,18 @@ function useSortFilterNodesButton({
         boolToInclude,
       ])
     }
-    const itemList = items.map((i) => Ast.TextLiteral.new(i))
+    const valueFormatter = getColumnValueToEnso(columnName)
+    if (items?.length === 1) {
+      const item = items[0]
+      if (item) {
+        return filterPattern.value.instantiateCopied([
+          Ast.TextLiteral.new(columnName),
+          Ast.parseExpression('..Equal')!,
+          valueFormatter(item, module) as Expression | MutableExpression,
+        ])
+      }
+    }
+    const itemList = items.map((i) => valueFormatter(i, module))
     return filterPattern.value.instantiateCopied([
       Ast.TextLiteral.new(columnName),
       Ast.parseExpression('..Is_In')!,
@@ -122,7 +138,7 @@ function useSortFilterNodesButton({
     const sortModelValue = toValue(sortModel)
     if (Object.keys(filterModelValue).length) {
       for (const [columnName, columnFilter] of Object.entries(filterModelValue)) {
-        const items = columnFilter.values.map((item) => `${item}`)
+        const items = columnFilter.values
         const filterPatterns =
           sortModelValue.length ?
             getAstPatternFilterAndSort(columnName, items)