+++ app/ydoc-shared/src/util/data/text.ts
@@ -1,15 +1,19 @@
+import * as iter from 'enso-common/src/utilities/data/iter'
 import diff from 'fast-diff'
 import { rangeEncloses, rangeLength, type SourceRange } from '../../yjsModel'
-import { Resumable } from './iterable'
 
 export type SourceRangeEdit = { range: SourceRange; insert: string }
 
 /** Given text and a set of `TextEdit`s, return the result of applying the edits to the text. */
-export function applyTextEdits(oldText: string, textEdits: SourceRangeEdit[]) {
-  textEdits.sort((a, b) => a.range[0] - b.range[0])
+export function applyTextEdits(
+  oldText: string,
+  textEdits: ReadonlyArray<Readonly<SourceRangeEdit>>,
+) {
+  const editsOrdered = [...textEdits]
+  editsOrdered.sort((a, b) => a.range[0] - b.range[0])
   let start = 0
   let newText = ''
-  for (const textEdit of textEdits) {
+  for (const textEdit of editsOrdered) {
     newText += oldText.slice(start, textEdit.range[0])
     newText += textEdit.insert
     start = textEdit.range[1]
@@ -70,8 +74,8 @@ export function offsetEdit(textEdit: SourceRangeEdit, offset: number): SourceRan
 export function applyTextEditsToSpans(textEdits: SourceRangeEdit[], spansBefore: SourceRange[]) {
   // Gather start and end points.
   const numerically = (a: number, b: number) => a - b
-  const starts = new Resumable(spansBefore.map(([start, _end]) => start).sort(numerically))
-  const ends = new Resumable(spansBefore.map(([_start, end]) => end).sort(numerically))
+  const starts = new iter.Resumable(spansBefore.map(([start, _end]) => start).sort(numerically))
+  const ends = new iter.Resumable(spansBefore.map(([_start, end]) => end).sort(numerically))
 
   // Construct translations from old locations to new locations for all start and end points.
   const startMap = new Map<number, number>()