+++ CHANGELOG.md
@@ -22,6 +22,18 @@
 - [Table Input Widget has now a limit of 256 cells.][11448]
 - [Added an error message screen displayed when viewing a deleted
   component.][11452]
+- [New documentation editor provides improved Markdown editing experience, and
+  paves the way for new documentation features.][11469]
+- [You can now add images to documentation panel][11547] by pasting them from
+  clipboard or by drag'n'dropping image files.
+- ["Write" button in component menu allows to evaluate it separately from the
+  rest of the workflow][11523].
+- [The documentation editor can now display tables][11564]
+- [The documentation editor supports the Markdown URL syntax, and uses it to
+  render pasted URLs as links][11597]
+- [Table Input Widget is now matched for Table.input method instead of
+  Table.new. Values must be string literals, and their content is parsed to the
+  suitable type][11612].
 
 [11151]: https://github.com/enso-org/enso/pull/11151
 [11271]: https://github.com/enso-org/enso/pull/11271
@@ -37,6 +49,12 @@
 [11447]: https://github.com/enso-org/enso/pull/11447
 [11448]: https://github.com/enso-org/enso/pull/11448
 [11452]: https://github.com/enso-org/enso/pull/11452
+[11469]: https://github.com/enso-org/enso/pull/11469
+[11547]: https://github.com/enso-org/enso/pull/11547
+[11523]: https://github.com/enso-org/enso/pull/11523
+[11564]: https://github.com/enso-org/enso/pull/11564
+[11597]: https://github.com/enso-org/enso/pull/11597
+[11612]: https://github.com/enso-org/enso/pull/11612
 
 #### Enso Standard Library
 
@@ -46,11 +64,17 @@
   programmatically.][11255]
 - [DB_Table may be saved as a Data Link.][11371]
 - [Support for dates before 1900 in Excel and signed AWS requests.][11373]
+- [Added `Data.read_many` that allows to read a list of files in a single
+  operation.][11490]
+- [Added `Table.input` allowing creation of typed tables from vectors of data,
+  including auto parsing text columns.][11562]
 
 [11235]: https://github.com/enso-org/enso/pull/11235
 [11255]: https://github.com/enso-org/enso/pull/11255
 [11371]: https://github.com/enso-org/enso/pull/11371
 [11373]: https://github.com/enso-org/enso/pull/11373
+[11490]: https://github.com/enso-org/enso/pull/11490
+[11562]: https://github.com/enso-org/enso/pull/11562
 
 #### Enso Language & Runtime